# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TourVision API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tours:
    get:
      operationId: listTours
      tags: [tours]
      summary: List user tours
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, ready, processing, failed]
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, confidence, views]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tours
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursListResponse"
    post:
      operationId: createTour
      tags: [tours]
      summary: Create a new tour from listing URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourBody"
      responses:
        "200":
          description: Tour created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTourResponse"
        "403":
          description: Tour limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tours/{tourId}:
    get:
      operationId: getTour
      tags: [tours]
      summary: Get a tour by ID (owner only)
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tour details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        "404":
          description: Tour not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTour
      tags: [tours]
      summary: Delete a tour
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tours/{tourId}/status:
    get:
      operationId: getTourStatus
      tags: [tours]
      summary: Get tour processing status (for polling)
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tour status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourStatus"
  /tours/{tourId}/rooms:
    get:
      operationId: getTourRooms
      tags: [tours]
      summary: Get rooms for a tour
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tour rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourRoomsResponse"
    put:
      operationId: updateTourRoomFloors
      tags: [tours]
      summary: Update floor assignments for tour rooms
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomFloorsBody"
      responses:
        "200":
          description: Rooms updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /tours/{tourId}/floors:
    put:
      operationId: setTourFloorCount
      tags: [tours]
      summary: Set number of floors for a tour
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetFloorCountBody"
      responses:
        "200":
          description: Floor count set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tours/public/{shareToken}:
    get:
      operationId: getPublicTour
      tags: [tours]
      summary: Get public tour by share token (no auth required)
      parameters:
        - name: shareToken
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public tour data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTour"
        "404":
          description: Tour not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    Tour:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        listingUrl:
          type: string
        listingTitle:
          type: string
          nullable: true
        listingAddress:
          type: string
          nullable: true
        listingPlatform:
          type: string
          nullable: true
        listingPrice:
          type: string
          nullable: true
        listingBedrooms:
          type: integer
          nullable: true
        listingBathrooms:
          type: integer
          nullable: true
        listingSqft:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, extracting, generating, stitching, building_3d, ready, failed]
        currentStage:
          type: string
          nullable: true
        totalPhotosExtracted:
          type: integer
        photosUsed:
          type: integer
        roomsDetected:
          type: integer
        floorCount:
          type: integer
          nullable: true
        confidenceScore:
          type: number
        realAngles:
          type: integer
        aiHighAngles:
          type: integer
        aiLowAngles:
          type: integer
        marbleWorldIds:
          type: object
          nullable: true
        tourEmbedUrl:
          type: string
          nullable: true
        shareToken:
          type: string
          nullable: true
        isWatermarked:
          type: boolean
        thumbnailUrl:
          type: string
          nullable: true
        viewCount:
          type: integer
        errorMessage:
          type: string
          nullable: true
        processingStartedAt:
          type: string
          nullable: true
        processingCompletedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - listingUrl
        - status
        - totalPhotosExtracted
        - photosUsed
        - roomsDetected
        - confidenceScore
        - realAngles
        - aiHighAngles
        - aiLowAngles
        - isWatermarked
        - viewCount
        - createdAt
    TourRoom:
      type: object
      properties:
        id:
          type: string
        tourId:
          type: string
        roomLabel:
          type: string
        floorNumber:
          type: integer
          nullable: true
        qualityScore:
          type: integer
        isSelected:
          type: boolean
        isBestForRoom:
          type: boolean
        confidenceScore:
          type: number
          nullable: true
        marbleWorldId:
          type: string
          nullable: true
        marbleEmbedUrl:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - tourId
        - roomLabel
        - qualityScore
        - isSelected
        - isBestForRoom
        - isAiGenerated
        - createdAt
    TourRoomsResponse:
      type: object
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/TourRoom"
        floorCount:
          type: integer
          nullable: true
      required:
        - rooms
    UpdateRoomFloorsBody:
      type: object
      properties:
        assignments:
          type: array
          items:
            type: object
            properties:
              roomId:
                type: string
              floorNumber:
                type: integer
            required:
              - roomId
              - floorNumber
      required:
        - assignments
    SetFloorCountBody:
      type: object
      properties:
        floorCount:
          type: integer
      required:
        - floorCount
    ToursListResponse:
      type: object
      properties:
        tours:
          type: array
          items:
            $ref: "#/components/schemas/Tour"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - tours
        - total
    CreateTourBody:
      type: object
      properties:
        listingUrl:
          type: string
        floorCount:
          type: integer
      required:
        - listingUrl
    CreateTourResponse:
      type: object
      properties:
        tourId:
          type: string
        shareToken:
          type: string
      required:
        - tourId
        - shareToken
    TourStatus:
      type: object
      properties:
        status:
          type: string
        currentStage:
          type: string
          nullable: true
        roomsCompleted:
          type: integer
        roomsTotal:
          type: integer
        estimatedMinutes:
          type: integer
        confidenceScore:
          type: number
        errorMessage:
          type: string
          nullable: true
      required:
        - status
        - roomsCompleted
        - roomsTotal
        - estimatedMinutes
        - confidenceScore
...
    PublicTour:
      type: object
      properties:
        id:
          type: string
        shareToken:
          type: string
        listingAddress:
          type: string
          nullable: true
        listingPlatform:
          type: string
          nullable: true
        listingPrice:
          type: string
          nullable: true
        listingBedrooms:
          type: integer
          nullable: true
        listingBathrooms:
          type: integer
          nullable: true
        listingSqft:
          type: string
          nullable: true
        marbleWorldIds:
          type: object
          nullable: true
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/TourRoom"
        confidenceScore:
          type: number
        realAngles:
          type: integer
        aiHighAngles:
          type: integer
        aiLowAngles:
          type: integer
        isWatermarked:
          type: boolean
        agentName:
          type: string
          nullable: true
        agentLogo:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
      required:
        - id
        - shareToken
        - confidenceScore
        - realAngles
        - aiHighAngles
        - aiLowAngles
        - isWatermarked
        - rooms
...